# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API documentation for Tatum Middleware. Tatum Middleware is a client to Tatum Blockchain API Core.",
    "version": "1.0.2",
    "title": "Tatum Middleware API"
  },
  "host": "localhost:6543",
  "basePath": "\/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "\/api\/v1\/withdrawal\/broadcast": {
      "post": {
        "tags": [
          "Withdrawal Endpoint"
        ],
        "summary": "Broadcast signed transaction and complete withdrawal",
        "description": "<p>Broadcast signed raw transaction end complete withdrawal associated with it. When broadcast succeeded but it is impossible to complete withdrawal, transaction id of transaction is returned and withdrawal must be completed manually.</p>",
        "operationId": "broadcast",
        "consumes": [
          "application\/json"
        ],
        "produces": [
          "application\/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "broadcast",
            "description": "broadcast",
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/Broadcast"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#\/definitions\/BroadcastResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ],
        "deprecated": false
      }
    },
...
    "\/api\/v1\/withdrawal\/{id}": {
      "delete": {
        "tags": [
          "Withdrawal Endpoint"
        ],
        "summary": "Cancel withdrawal",
        "description": "<p>This method is helpful if you need to cancel withdrawal in case of blockchain transaction failed or is not yet processed. This does not cancel already broadcast blockchain transaction, only Tatum internal withdrawal request.</p>",
        "operationId": "cancelWithdrawal",
        "produces": [
          "*\/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of created withdrawal",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ],
        "deprecated": false
      }
    },
...
    "\/util\/v1\/btc\/wallet": {
      "post": {
        "parameters": [
          {
            "name": "chain",
            "in": "body",
            "description": "chain - BTC for Mainnet or TBTC for Testnet3",
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/WalletGenerateBtc"
            }
          }
        ],
        "operationId": "generateBtcWallet",
        "summary": "Generate wallet.",
        "description": "<p>Generate wallet. This will create BIP44 HD wallet with predefined derivation path and returns mnemonic phrase and extended public and private key to be used for address generation.</p>",
        "tags": [
          "Bitcoin blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Object containing mnemonic, xpriv and xpub for generated wallet.",
            "schema": {
              "$ref": "#\/definitions\/Wallet"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/btc\/wallet\/xpub\/{chain}\/{xpub}\/{i}": {
      "get": {
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "chain - BTC for Mainnet or TBTC for Testnet3",
            "required": true,
            "type": "string",
            "enum": [
              "BTC",
              "TBTC"
            ]
          },
          {
            "name": "xpub",
            "in": "path",
            "description": "xpub to generate address from",
            "required": true,
            "type": "string"
          },
          {
            "name": "i",
            "in": "path",
            "description": "derivation index of address",
            "required": true,
            "type": "integer"
          }
        ],
        "operationId": "calculateBtcAddress",
        "summary": "Calculate BTC address",
        "description": "<p>Calculate address from extended public key on Testnet \/ Mainnet for given derivation path index.</p>",
        "tags": [
          "Bitcoin blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Generated address",
            "schema": {
              "$ref": "#\/definitions\/GeneratedAddress"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/btc\/wallet\/xpriv": {
      "post": {
        "parameters": [
          {
            "name": "xpriv",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/XPrivBtc"
            }
          }
        ],
        "operationId": "calculateBtcPrviKey",
        "summary": "Calculate BTC private key",
        "description": "<p>Calculate private key of address from mnemonic on Testnet \/ Mainnet for given derivation path index.</p>",
        "tags": [
          "Bitcoin blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "Generated private key",
            "schema": {
              "$ref": "#\/definitions\/PrivKey"
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        }
      }
    },
    "\/util\/v1\/btc\/withdrawal": {
      "post": {
        "parameters": [
          {
            "name": "transfer",
            "in": "body",
            "description": null,
            "required": true,
            "schema": {
              "$ref": "#\/definitions\/WithdrawalBtc"
            }
          }
        ],
        "operationId": "transferBtc",
        "summary": "Send BTC \/ TBTC from address to address",
        "description": "<p>Send BTC \/ TBTC from address to address. This will create Tatum internal withdrawal request with ID. If every system works as expected, withdrawal request is marked as complete and transaction id is assigned to it.<ul><li>If Bitcoin server connection is unavailable, withdrawal request is cancelled.</li><li>If blockchain transfer is successful, but is it not possible to reach Tatum Blockchain API, transaction id of blockchain transaction is returned and withdrawal request must be completed manually, otherwise all other withdrawals will be pending.</li></ul></p>",
        "tags": [
          "Bitcoin blockchain helper Endpoint"
        ],
        "responses": {
          "200": {
            "description": "txHash of successful transaction",
            "schema": {
              "$ref": "#\/definitions\/TxHash"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          },
          "500": {
            "description": "Internal error",
            "schema": {
              "$ref": "#\/definitions\/Error"
            }
          }
        },
        "security": [
          {
            "apiKey": [
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "JWT token generated from `apiKey` field of your API Key and current timestamp signed by `secret` of your API Key passed as Bearer token in authorization header, e.g. `authorization: Bearer 12kjn128n123n12i3n`.<br/>`{\"apiKey\": \"your_api_key\", \"created\": current_timestamp_millis}`<br/>How to create valid **JWT** can be found at <a href=\"https://jwt.io\">https://jwt.io</a>, or you can use our JWT helper Endpoint."
    }
  },
...
  "produces": [
    "application\/json"
  ],
  "consumes": [
    "application\/json"
  ]
}